G can be double

+++ treewalk.h
@@ -10,6 +10,6 @@ double PotentialKernel(double r, double h);
 double ForceKernel(double r, double h);
 
 double PotentialWalk_quad(double* pos, Octree* tree, double theta, double softening);
-double* PotentialTarget_tree(int Npar, double** pos_target, double* softening_target, Octree* tree, int G, double theta);
+double* PotentialTarget_tree(int Npar, double** pos_target, double* softening_target, Octree* tree, double G, double theta);
 double* AccelWalk_quad(double* pos, Octree* tree, double theta, double softening);
 double** AccelTarget_tree(int Npar, double** pos_target, double* softening_target, Octree* tree, double G, double theta);
